+++ src/PowerSimulations.jl
@@ -434,7 +434,6 @@ using DocStringExtensions
 include("core/definitions.jl")
 
 # Core components
-include("core/abstract_formulations.jl")
 include("core/formulations.jl")
 include("core/abstract_simulation_store.jl")
 include("core/operation_model_abstract_types.jl")
